+++ internal/gps/strip_vendor.go
@@ -16,25 +16,23 @@ func stripVendor(path string, info os.FileInfo, err error) error {
 		return err
 	}
 
-	if info.Name() == "vendor" {
-		if _, err := os.Lstat(path); err != nil {
+	// Skip anything not named vendor
+	if info.Name() != "vendor" {
+		return nil
+	}
+
+  // If the file is a symlink to a directory, delete the symlink.
+  if (info.Mode() & os.ModeSymlink) != 0 {
+		if realInfo, err := os.Stat(path); err == nil && realInfo.IsDir() {
+			return os.Remove(path)
+		}
+	}
+
+	if info.IsDir() {
+		if err := os.RemoveAll(path); err != nil {
 			return err
 		}
-
-		if (info.Mode() & os.ModeSymlink) != 0 {
-			if realInfo, err := os.Stat(path); err == nil && realInfo.IsDir() {
-				return os.Remove(path)
-			}
-		}
-
-		if info.IsDir() {
-			if err := os.RemoveAll(path); err != nil {
-				return err
-			}
-			return filepath.SkipDir
-		}
-
-		return nil
+		return filepath.SkipDir
 	}
 
 	return nil