+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # v0.3.2 (Unreleased)
 
+BUG FIXES:
+
+* Fix `status` shows incorrect reason for lock mismatch when ignoring packages.
+(#1216)
+
+IMPROVEMENTS:
+
+* `dep ensure -add` now concurrently fetches the source and adds the projects.
+(#1218)
+* File name case check is now performed on `Gopkg.toml` and `Gopkg.lock`.
+(#1114)
+* gps: gps now supports pruning. (#1020)
+* `dep ensure -update` now concurrently validates the passed project arguments.
+Improving performance when updating dependencies with `-update`. (#1175)
+* `dep status` now concurrently fetches repo info. Improving status performance.
+(#1135)
+
+
 # v0.3.1
 
 * gps: Add satisfiability check for case variants (#1079)